Removing wrapper and passing methods to DbConnectionExtensions
