+++ src/Take.Elephant.Sql/SqlMap.cs
@@ -16,12 +16,9 @@ namespace Take.Elephant.Sql
     /// <typeparam name="TValue"></typeparam>
     public class SqlMap<TKey, TValue> : MapStorageBase<TKey, TValue>, IKeysMap<TKey, TValue>, IPropertyMap<TKey, TValue>, IUpdatableMap<TKey, TValue>
     {
-        private readonly IDbConnectionExtensionInstatiable _dbConnectionExtensionInstatiable;
-
-        public SqlMap(string connectionString, ITable table, IMapper<TKey> keyMapper, IMapper<TValue> valueMapper,IDbConnectionExtensionInstatiable dbConnectionExtensionInstatiable)
+        public SqlMap(string connectionString, ITable table, IMapper<TKey> keyMapper, IMapper<TValue> valueMapper)
             : this(new SqlDatabaseDriver(), connectionString, table, keyMapper, valueMapper)
         {
-            _dbConnectionExtensionInstatiable = dbConnectionExtensionInstatiable;
         }
 
         public SqlMap(IDatabaseDriver databaseDriver, string connectionString, ITable table, IMapper<TKey> keyMapper, IMapper<TValue> valueMapper)
@@ -165,7 +162,7 @@ namespace Take.Elephant.Sql
                 {
                     var keyColumnValues = KeyMapper.GetColumnValues(key);
 
-                    using (var command = _dbConnectionExtensionInstatiable.CreateSelectTop1Command(connection, DatabaseDriver, Table, new[] { propertyName }, keyColumnValues))
+                    using (var command = ConnectionExtensions.CreateSelectTop1Command(connection, DatabaseDriver, Table, new[] { propertyName }, keyColumnValues))
                     {
                         var dbValue = await command.ExecuteScalarAsync(cancellationTokenSource.Token).ConfigureAwait(false);
                         if (dbValue != null && !(dbValue is DBNull))
